Fix some pointer warnings
Fix lots of the following warnings:

"assignment from incompatible pointer type"
"assignment makes integer from pointer without a cast"
"assignment makes pointer from integer without a cast"

but still sooooo many warnings to go....

+++ arch/arm/mach-tcc88xx/tca_cipher.c
@@ -296,9 +296,9 @@ void tca_cipher_set_baseaddr(unsigned uTxRx, unsigned char *pBaseAddr)
 	dprintk("%s\n", __func__);
 
 	if(uTxRx == TCC_CIPHER_BASEADDR_TX)
-		pHwCIPHER->TXBASE = pBaseAddr;
+		pHwCIPHER->TXBASE = (unsigned int)pBaseAddr;
 	else
-		pHwCIPHER->RXBASE = pBaseAddr;
+		pHwCIPHER->RXBASE = (unsigned int)pBaseAddr;
 }
 
 void tca_cipher_set_packet(unsigned uCount, unsigned uSize)
